refactor(review): 코드 리뷰 피드백 반영 — dead code 제거 및 도메인 간 의존성 해결
Gemini 코드 리뷰 #2/3/6 반영.

- ReviewItemBlock의 미사용 gap/icon prop과 주석 처리 코드 제거
- line-clamp 분기 9-단계 ternary를 객체 매핑으로 개선 (가독성)
- ReviewCard의 questionIcon 호출 제거 (icon prop 사용처 dead chain 정리)
- reviewCardHelpers.tsx → .ts (JSX 헬퍼 제거로 SVG import 7개 정리)
- domain/review/form/ → common/review-form/ 이동 (3+ 도메인 사용 — Rule 59)
  - GoalOrConcernsBox, ReviewInstruction, ReviewQuestion, ReviewTextarea
  - mypage 6페이지 + challenge 4파일 import 수정
