+++ apps/web/src/app/(user)/mypage/review/new/live/[programId]/page.tsx
@@ -8,10 +8,10 @@ import { Suspense, useState } from 'react';
 
 import { useGetLiveTitle } from '@/api/program';
 import { usePostReviewMutation } from '@/api/review/review';
-import ReviewInstruction from '@/domain/review/form/ReviewInstruction';
+import ReviewInstruction from '@/common/review-form/ReviewInstruction';
 import ReviewModal from '@/domain/review/modal/ReviewModal';
-import ReviewQuestion from '@/domain/review/form/ReviewQuestion';
-import ReviewTextarea from '@/domain/review/form/ReviewTextarea';
+import ReviewQuestion from '@/common/review-form/ReviewQuestion';
+import ReviewTextarea from '@/common/review-form/ReviewTextarea';
 import TenScore from '@/domain/review/score/TenScore';
 
 const LiveReviewCreatePageContent = () => {
