+++ apps/web/src/domain/review/ui/ReviewItemBlock.tsx
@@ -1,7 +1,6 @@
 import { QuestionType } from '@/api/review/review';
 import { twMerge } from '@/lib/twMerge';
 import { questionTypeToText } from '@/utils/convert';
-import { ReactNode } from 'react';
 import ExpandableParagraph from './ExpandableParagraph';
 
 interface Props {
@@ -11,9 +10,15 @@ interface Props {
   /** 이게 있을 경우 우선함. */
   questionText?: string;
   expandable?: boolean;
-  icon?: ReactNode;
-  gap?: 'normal' | 'large';
 }
+
+const LINE_CLAMP_CLASS = {
+  1: 'line-clamp-1',
+  2: 'line-clamp-2',
+  3: 'line-clamp-3',
+  4: 'line-clamp-4',
+  5: 'line-clamp-5',
+} as const;
 
 const ReviewItemBlock = (props: Props) => {
   const questionText =
@@ -23,7 +28,6 @@ const ReviewItemBlock = (props: Props) => {
   return (
     <div>
       <div className="flex w-fit items-center gap-1">
-        {/* {props.icon && props.icon} */}
         <span className="text-xsmall14 text-neutral-10 font-semibold">
           {questionText}
         </span>
@@ -38,17 +42,7 @@ const ReviewItemBlock = (props: Props) => {
         <p
           className={twMerge(
             'text-xsmall14 text-neutral-20 font-normal tracking-[-0.028px]',
-            props.lineClamp === 1
-              ? 'line-clamp-1'
-              : props.lineClamp === 2
-                ? 'line-clamp-2'
-                : props.lineClamp === 3
-                  ? 'line-clamp-3'
-                  : props.lineClamp === 4
-                    ? 'line-clamp-4'
-                    : props.lineClamp === 5
-                      ? 'line-clamp-5'
-                      : null,
+            props.lineClamp ? LINE_CLAMP_CLASS[props.lineClamp] : '',
           )}
         >
           {props.answer}
